+++ docsig/_module.py
@@ -8,15 +8,19 @@ from pathlib import Path as _Path
 
 import astroid as _ast
 
+from ._disable import Disabled as _Disabled
 from ._function import Function as _Function
 from ._objects import MutableSequence as _MutableSequence
 from ._utils import isprotected as _isprotected
 
 
-class Parent(_MutableSequence[_Function]):
+class Parent(  # pylint: disable=too-many-arguments
+    _MutableSequence[_Function]
+):
     """Represents an object that contains functions or methods.
 
     :param node: Parent's abstract syntax tree.
+    :param disabled: Data for lines which are excluded from checks.
     :param path: Path to base path representation on.
     :param ignore_args: Ignore args prefixed with an asterisk.
     :param ignore_kwargs: Ignore kwargs prefixed with two asterisks.
@@ -25,6 +29,7 @@ class Parent(_MutableSequence[_Function]):
     def __init__(
         self,
         node: _ast.Module | _ast.ClassDef,
+        disabled: _Disabled,
         path: _Path | None = None,
         ignore_args: bool = False,
         ignore_kwargs: bool = False,
@@ -32,18 +37,29 @@ class Parent(_MutableSequence[_Function]):
         super().__init__()
         self._name = node.name
         self._path = f"{path}:" if path is not None else ""
+        self._directives = disabled
         overloads = []
         returns = None
         for subnode in node.body:
             if isinstance(subnode, _ast.FunctionDef):
-                func = _Function(subnode, ignore_args, ignore_kwargs)
+                func = _Function(
+                    subnode,
+                    disabled.get(subnode.lineno, []),
+                    ignore_args,
+                    ignore_kwargs,
+                )
                 if func.isoverloaded:
                     overloads.append(func.name)
                     returns = func.signature.rettype
                 else:
                     if func.name in overloads:
                         subnode.returns = returns
-                        func = _Function(subnode, ignore_kwargs, ignore_kwargs)
+                        func = _Function(
+                            subnode,
+                            disabled.get(subnode.lineno, []),
+                            ignore_args,
+                            ignore_kwargs,
+                        )
 
                     self.append(func)
 
@@ -68,10 +84,13 @@ class _Module(_MutableSequence[Parent]):
     ) -> None:
         super().__init__()
         ast = _ast.parse(string)
-        self.append(Parent(ast, path, ignore_args, ignore_kwargs))
+        disabled = _Disabled(string)
+        self.append(Parent(ast, disabled, path, ignore_args, ignore_kwargs))
         for subnode in ast.body:
             if isinstance(subnode, _ast.ClassDef):
-                self.append(Parent(subnode, path, ignore_args, ignore_kwargs))
+                self.append(
+                    Parent(subnode, disabled, path, ignore_args, ignore_kwargs)
+                )
 
 
 class Modules(_MutableSequence[_Module]):