+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 [Unreleased](https://github.com/jshwi/docsig/compare/v0.35.0...HEAD)
 ------------------------------------------------------------------------
 ### Added
+- add support for disable comments
 - add support for overloaded functions
 
 [0.35.0](https://github.com/jshwi/docsig/releases/tag/v0.35.0) - 2023-08-09